+++ react/StoreLink.tsx
@@ -93,11 +93,6 @@ function StoreLink(props: Props) {
     intl,
   })
 
-  const handlePrevent = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    e.nativeEvent.stopImmediatePropagation()
-  }
-
   return (
     <Link
       to={resolvedLink}
@@ -105,7 +100,6 @@ function StoreLink(props: Props) {
       className={rootClasses}
       scrollOptions={scrollOptions}
       rel={rel}
-      onClick={handlePrevent}
     >
       {label && <span className={labelClasses}>{localizedLabel}</span>}
       {hasChildren(children) && (